+++ elasticapm/contrib/opentracing/__init__.py
@@ -1,0 +1,2 @@
+from .span import OTSpan  # noqa: F401
+from .tracer import Tracer  # noqa: F401